requirements.txt
python-dotenv
python-telegram-bot
tinydb
more_itertools
